+++ src/lib.rs
@@ -40,18 +40,24 @@ pub mod slave;
 pub use self::slave::{Slave, SlaveId};
 
 mod codec;
+
 mod error;
+pub use self::error::{Error, ProtocolError};
 
 mod frame;
-pub use self::frame::{Address, Exception, FunctionCode, Quantity, Request, Response};
+pub use self::frame::{
+    Address, Exception, ExceptionResponse, FunctionCode, Quantity, Request, Response,
+};
 
-/// Specialized [`std::result::Result`] type for `Modbus` client API.
+/// Specialized [`std::result::Result`] type for type-checked responses of the _Modbus_ client API.
+///
+/// The payload is generic over the response type.
 ///
 /// This [`Result`] type contains 2 layers of errors.
 ///
-/// 1. [`std::io::Error`]: An error occurred while performing I/O operations.
-/// 2. [`Exception`]: An error occurred on the `Modbus` server.
-pub type Result<T> = std::io::Result<std::result::Result<T, Exception>>;
+/// 1. [`Error`]: An unexpected protocol or network error that occurred during client/server communication.
+/// 2. [`Exception`]: An error occurred on the _Modbus_ server.
+pub type Result<T> = std::result::Result<std::result::Result<T, Exception>, Error>;
 
 mod service;