+++ tests/exception/mod.rs
@@ -42,66 +42,70 @@ impl Service for TestService {
 
 // TODO: Update the `assert_eq` with a check on Exception once Client trait can return Exception
 pub async fn check_client_context(mut ctx: Context) {
-    let response = ctx.read_coils(0x00, 2).await.expect("communication failed");
-    assert_eq!(response, Err(Exception::Acknowledge));
+    let response = ctx.read_coils(0x00, 2).await.unwrap();
+    assert!(matches!(response, Err(Exception::Acknowledge)));
 
     let response = ctx
         .read_discrete_inputs(0x00, 2)
         .await
         .expect("communication failed");
-    assert_eq!(response, Err(Exception::GatewayPathUnavailable));
+    assert!(matches!(response, Err(Exception::GatewayPathUnavailable)));
 
     let response = ctx
         .write_single_coil(0x00, true)
         .await
         .expect("communication failed");
-    assert_eq!(response, Err(Exception::GatewayTargetDevice));
+    assert!(matches!(response, Err(Exception::GatewayTargetDevice)));
 
     let response = ctx
         .write_multiple_coils(0x00, &[true])
         .await
         .expect("communication failed");
-    assert_eq!(response, Err(Exception::IllegalDataAddress));
+    assert!(matches!(response, Err(Exception::IllegalDataAddress)));
 
     let response = ctx
         .read_input_registers(0x00, 2)
         .await
         .expect("communication failed");
-    assert_eq!(response, Err(Exception::IllegalDataValue));
+    assert!(matches!(response, Err(Exception::IllegalDataValue)));
 
     let response = ctx
         .read_holding_registers(0x00, 2)
         .await
         .expect("communication failed");
-    assert_eq!(response, Err(Exception::IllegalFunction));
+    assert!(matches!(response, Err(Exception::IllegalFunction)));
 
     let response = ctx
         .write_single_register(0x00, 42)
         .await
         .expect("communication failed");
-    assert_eq!(response, Err(Exception::MemoryParityError));
+    assert!(matches!(response, Err(Exception::MemoryParityError)));
 
     let response = ctx
         .write_multiple_registers(0x00, &[42])
         .await
         .expect("communication failed");
-    assert_eq!(response, Err(Exception::ServerDeviceBusy));
+    assert!(matches!(response, Err(Exception::ServerDeviceBusy)));
 
     let response = ctx
         .masked_write_register(0x00, 0, 0)
         .await
         .expect("communication failed");
-    assert_eq!(response, Err(Exception::ServerDeviceFailure));
+    assert!(matches!(response, Err(Exception::ServerDeviceFailure)));
 
     let response = ctx
         .read_write_multiple_registers(0x00, 0, 0, &[42])
         .await
         .expect("communication failed");
-    assert_eq!(response, Err(Exception::IllegalFunction));
+    assert!(matches!(response, Err(Exception::IllegalFunction)));
 
     // TODO: This codes hangs if used with `rtu-over-tcp-server`, need to check why
-    /*
-    let response = ctx.call(Request::Custom(70, Cow::Owned(vec![42]))).await.expect("communication failed");
-    assert_eq!(response, Err(Exception::IllegalFunction));
-    */
+    // let response = ctx
+    //     .call(Request::Custom(70, Cow::Owned(vec![42])))
+    //     .await
+    //     .expect("communication failed");
+    // assert!(matches!(
+    //     response,
+    //     Err(Exception::IllegalFunction)
+    // ));
 }